move code to ghost like directory structure
