+++ trunk/gl_rmain.c
@@ -3585,122 +3585,6 @@ void R_DrawTransEntitiesOnList()
 	}
 }
 
-void R_DrawPathTracer(void)
-{
-	if (sv_player == NULL) return;
-	if (!sv.active && !cls.demoplayback) return;
-	if (cls.demoplayback) return;
-
-	extern pathtracer_bunnyhop_t pathtracer_bunnyhop_samples[100];
-	extern pathtracer_movement_t pathtracer_movement_samples[100000];
-	extern float drawbestangle;
-
-	glPolygonMode(GL_FRONT_AND_BACK, GL_LINE);
-	glDisable(GL_TEXTURE_2D);
-	glDisable(GL_CULL_FACE);
-
-	glBegin(GL_LINES);
-	for (int i = 1;i < 100000;i++) {
-		pathtracer_movement_t* pms_prev = &pathtracer_movement_samples[i - 1];
-		pathtracer_movement_t* pms_cur = &pathtracer_movement_samples[i];
-
-		if (pms_prev->pos[0] == 0.f && pms_prev->pos[1] == 0.f && pms_prev->pos[2] == 0.f) {
-			continue;
-		}
-		if (pms_cur->pos[0] == 0.f && pms_cur->pos[1] == 0.f && pms_cur->pos[2] == 0.f) {
-			continue;
-		}
-
-		if (pms_cur->ongound && pms_prev->ongound) {
-			glColor3f(.3f, 0.f, 0.f);
-		}
-		else if (pms_cur->ongound || pms_prev->ongound) {
-			glColor3f(.1f, .1f, .1f);
-		}
-		else {
-			glColor3f(1.f, 1.f, 1.f);
-		}
-
-		GLfloat startPos[3];
-		startPos[0] = pms_prev->pos[0];
-		startPos[1] = pms_prev->pos[1];
-		startPos[2] = pms_prev->pos[2];// -20.f;
-		glVertex3fv(startPos);
-
-		GLfloat startPos2[3];
-		startPos2[0] = pms_cur->pos[0];
-		startPos2[1] = pms_cur->pos[1];
-		startPos2[2] = pms_cur->pos[2];// -20.f;
-		glVertex3fv(startPos2);
-	}
-	glEnd();
-
-	for (int i = 1;i < 100000;i++) {
-		pathtracer_movement_t* pms_prev = &pathtracer_movement_samples[i - 1];
-		pathtracer_movement_t* pms_cur = &pathtracer_movement_samples[i];
-
-		if (pms_prev->pos[0] == 0.f && pms_prev->pos[1] == 0.f && pms_prev->pos[2] == 0.f) {
-			continue;
-		}
-		if (pms_cur->pos[0] == 0.f && pms_cur->pos[1] == 0.f && pms_cur->pos[2] == 0.f) {
-			continue;
-		}
-
-		vec3_t playerforward, right, up;
-		vec3_t playerangles = { 0.f,pms_cur->angle ,0.f };
-
-		AngleVectors(playerangles, playerforward, right, up);
-
-		float playerSpeed = (pms_cur->speed - pms_prev->speed) * 2.f;
-		float bestSpeed = 0; // pms_cur->bestspeed * 10.f;
-
-		float oldspeed = sqrtf(pms_prev->velocity[0] * pms_prev->velocity[0] + pms_prev->velocity[1] * pms_prev->velocity[1]);
-		float newspeed = sqrtf(pms_cur->velocity[0] * pms_cur->velocity[0] + pms_cur->velocity[1] * pms_cur->velocity[1]);
-		float dspeed = newspeed - oldspeed;
-
-		// velocity change, green faster, red slower
-		glBegin(GL_LINES);
-		if (dspeed > 0)
-			glColor3f(0.f, 1.f, 0.f);
-		else if (dspeed == 0)
-			glColor3f(1.f, 1.f, 1.f);
-		else
-			glColor3f(1.f, 0.f, 0.f);
-		GLfloat startPos[3];
-		startPos[0] = pms_cur->pos[0];
-		startPos[1] = pms_cur->pos[1];
-		startPos[2] = pms_cur->pos[2];// -20.f;
-		glVertex3fv(startPos);
-
-		GLfloat anglePos[3];
-		anglePos[0] = startPos[0] + (pms_cur->velocity[0] - pms_prev->velocity[0]); //playerforward[0] * playerSpeed;
-		anglePos[1] = startPos[1] + (pms_cur->velocity[1] - pms_prev->velocity[1]); //playerforward[1] * playerSpeed;
-		anglePos[2] = startPos[2];
-		glVertex3fv(anglePos);
-		glEnd();
-
-		// speed line in blue
-		glBegin(GL_LINES);
-		glColor3f(0.f, 0.f, 1.f);
-		startPos[0] = pms_prev->pos[0] + up[0] * (oldspeed - 300) / 3.f;
-		startPos[1] = pms_prev->pos[1] + up[1] * (oldspeed - 300) / 3.f;
-		startPos[2] = 0.f + up[2] * (oldspeed - 300) / 3.f;
-		glVertex3fv(startPos);
-
-		anglePos[0] = pms_cur->pos[0] + up[0] * (newspeed - 300) / 3.f;
-		anglePos[1] = pms_cur->pos[1] + up[1] * (newspeed - 300) / 3.f;
-		anglePos[2] = 0.f + up[2] * (newspeed - 300) / 3.f;
-		glVertex3fv(anglePos);
-		glEnd();
-
-	}
-
-	glColor3f(1, 1, 1);
-	glEnable(GL_TEXTURE_2D);
-	glEnable(GL_CULL_FACE);
-	glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
-}
-
 /*
 =============
 R_DrawViewModel
