+++ trunk/pathtracer/pathtracer_private.h
@@ -1,0 +1,8 @@
+#ifndef __PATHTRACER_PRIVATE
+#define __PATHTRACER_PRIVATE
+
+#ifndef QUAKE_GAME
+#include "../quakedef.h"
+#endif
+
+#endif // __PATHTRACER_PRIVATE
