+++ trunk/quakedef.h
@@ -253,6 +253,7 @@ typedef struct
 #include "iplog.h"
 #include "cl_slist.h"
 #include "ghost/ghost.h"
+#include "pathtracer/pathtracer.h"
 
 //=============================================================================
