+++ trunk/view.c
@@ -1387,255 +1387,6 @@ float Get_Wishdir_Speed_Delta(float angle) {
 	return newspeed - oldspeed;
 }
 
-
-pathtracer_bunnyhop_t pathtracer_bunnyhop_samples[100];
-pathtracer_movement_t pathtracer_movement_samples[100000];
-
-float drawbestangle = NAN;
-extern	cvar_t	scr_printbunnyhop;
-extern	cvar_t	scr_recordbunnyhop;
-
-
-void SCR_DrawSpeedHelp(void) {
-	// draw speed info for each jump
-	if (!sv.active && !cls.demoplayback) return;
-	if (!sv.active) return;
-	if (cls.demoplayback) return;
-	if (sv_player == NULL) return;
-
-	const char* fmt = "%c%c %3.0f %c %3.0f = %3.0f %c%3.0f %2.0f%% %6s";
-	char str[100];
-	int fmtlen = 1 + sprintf(str, fmt, ' ', ' ', 0.0, '+', 0.0, 0.0, '+', 0.0, 0.0, "");
-	int rows = (int)10;
-	if (rows <= 0) return;
-	if (rows > 100) rows = 100;
-
-	float s = fmin(20, fmax(1, 1));
-	int w = fmtlen * 8;
-	int h = rows * 8;
-
-	extern usercmd_t cmd;
-
-	static qboolean prevonground;
-	static qboolean prevprevonground;
-	static float prevspeed = 0;
-	static float airspeedsum = 0;
-	static float airspeedsummax = 0;
-	static double prevcltime = -1;
-	static int pathtracer_bunnyhop_index = 0;
-	static int pathtracer_movement_index = 0;
-	static int skipframe = 0;
-	static int skipframemod = 10;
-	static float currentclosestbunnyhopDistance = INFINITY;
-	static int currentclosestbunnyhopIndex = -1;
-	extern entity_t ghost_entity;
-
-
-	// demo playback only has cl.onground which is different from onground in normal play
-	qboolean onground0;
-	if (sv.active) {
-		extern qboolean onground;
-		onground0 = onground;
-	}
-	else {
-		onground0 = cl.onground;
-	}
-
-	// restarted
-	if (cl.time < prevcltime) {
-		memset(pathtracer_movement_samples, 0, sizeof(pathtracer_movement_samples));
-		pathtracer_movement_index = 0;
-
-		memset(pathtracer_bunnyhop_samples, 0, sizeof(pathtracer_bunnyhop_samples));
-		pathtracer_bunnyhop_index = 0;
-
-		currentclosestbunnyhopIndex = -1;
-		currentclosestbunnyhopDistance = INFINITY;
-		drawbestangle = NAN;
-
-		prevcltime = cl.time;
-	}
-	else {
-		prevcltime = cl.time;
-	}
-
-	{
-		int angles = 90;
-		float bestspeed = NAN;
-		float bestangle = 0;
-		bestangle = sv_player->v.angles[1];
-		bestspeed = Get_Wishdir_Speed_Delta(bestangle);
-		float playerspeed = Get_Wishdir_Speed_Delta(bestangle);
-		for (int i = 1; i < angles; i++) {
-			for (int neg = -1; neg <= 1; neg += 2) {
-				float curangle = 0.0f;
-				curangle = sv_player->v.angles[1] + i * neg;
-				float curspeed = Get_Wishdir_Speed_Delta(curangle);
-				if (curspeed > bestspeed) {
-					bestspeed = curspeed;
-					bestangle = curangle;
-				}
-			}
-		}
-		for (int i = 1; i < 100; i++) {
-			for (int neg = -1; neg <= 1; neg += 2) {
-				float curangle = bestangle + neg * i / 100.0f;
-				float curspeed = Get_Wishdir_Speed_Delta(curangle);
-				if (curspeed > (bestspeed)) {
-					bestspeed = curspeed;
-					bestangle = curangle;
-				}
-			}
-		}
-		if (!onground0) {
-			airspeedsum += Get_Wishdir_Speed_Delta(sv_player->v.angles[1]);
-			airspeedsummax += bestspeed;
-		}
-
-		if (bestangle < -180) bestangle += 360;
-		if (bestangle > 180) bestangle -= 360;
-
-
-		float speed = sqrt(cl.velocity[0] * cl.velocity[0] + cl.velocity[1] * cl.velocity[1]);
-
-		boolean track = false;
-		if (pathtracer_movement_index > 0) {
-			pathtracer_movement_t* pms_prev = &pathtracer_movement_samples[pathtracer_movement_index - 1];
-			float dx = fabs(pms_prev->pos[0] - sv_player->v.origin[0]);
-			float dy = fabs(pms_prev->pos[1] - sv_player->v.origin[1]);
-			float dz = fabs(pms_prev->pos[2] - sv_player->v.origin[2]);
-			if (dx > .1f || dy > .1f || dz > .1f) {
-				track = true;
-			}
-		}
-		else {
-			track = true;
-		}
-		if (track && scr_recordbunnyhop.value == 1.f) {
-			pathtracer_movement_t* pms_new = &pathtracer_movement_samples[pathtracer_movement_index];
-			pms_new->pos[0] = sv_player->v.origin[0];
-			pms_new->pos[1] = sv_player->v.origin[1];
-			pms_new->pos[2] = sv_player->v.origin[2];
-			pms_new->velocity[0] = sv_player->v.velocity[0];
-			pms_new->velocity[1] = sv_player->v.velocity[1];
-			pms_new->velocity[2] = sv_player->v.velocity[2];
-			pms_new->angle = sv_player->v.angles[1];
-
-			if (ghost_entity.model != NULL) {
-				pms_new->pos[0] = ghost_entity.origin[0];
-				pms_new->pos[1] = ghost_entity.origin[1];
-				pms_new->pos[2] = ghost_entity.origin[2];
-				pms_new->velocity[0] = ghost_entity.origin[0] - ghost_entity.previousorigin[0];
-				pms_new->velocity[1] = ghost_entity.origin[1] - ghost_entity.previousorigin[1];
-				pms_new->velocity[2] = ghost_entity.origin[2] - ghost_entity.previousorigin[2];
-				pms_new->velocity[0] *= 200.f;
-				pms_new->velocity[1] *= 200.f;
-				pms_new->velocity[2] *= 200.f;
-				pms_new->angle = ghost_entity.angles[1];
-			}
-
-			pms_new->bestangle = bestangle;
-			pms_new->speed = speed;
-			pms_new->bestspeed = bestspeed;
-			pms_new->ongound = onground0;
-			pathtracer_movement_index++;
-			if (pathtracer_movement_index >= 100000) {
-				pathtracer_movement_index = 0;
-			}
-		}
-		if ((onground0 && !prevonground) ||  // landed
-			(!onground0 && prevonground && prevprevonground)) { // take-off
-			char strextra[10];
-			float addspeed;
-			float addspeedpct;
-			strextra[0] = 0;
-			if (sv.active) {
-				addspeed = Get_Wishdir_Speed_Delta(sv_player->v.angles[1]);
-				addspeedpct = 100.0 * (airspeedsum + addspeed) / (airspeedsummax + bestspeed);
-				sprintf(strextra, "%4.0f %1s", fmin(999, fmax(-999, bestangle - sv_player->v.angles[1])), (onground0 && !prevonground) ? "l" : "t");
-			}
-			else { // demo
-				addspeed = 0.0;
-				addspeedpct = 0.0;
-				airspeedsum = 0.0;
-				addspeedpct = 0.0;
-			}
-			sprintf(pathtracer_bunnyhop_samples[pathtracer_bunnyhop_index].str, fmt,
-				cmd.sidemove < 0 ? 127 : cmd.sidemove > 0 ? 141 : ' ',
-				cmd.forwardmove > 0 ? '^' : cmd.forwardmove < 0 ? 'v' : ' ',
-				airspeedsum,
-				addspeed >= 0 ? '+' : '-',
-				fabs(addspeed),
-				fmin(999.0, speed),
-				speed - prevspeed >= 0 ? '+' : '-',
-				fmin(999.0, fabs(speed - prevspeed)),
-				fmin(99, fmax(0, addspeedpct)),
-				strextra);
-
-			prevspeed = speed;
-			airspeedsum = 0;
-			airspeedsummax = 0;
-
-			if (onground0 && !prevonground && scr_recordbunnyhop.value == 1.f) {
-				drawbestangle = bestangle;
-				pathtracer_bunnyhop_samples[pathtracer_bunnyhop_index].playerangle = sv_player->v.angles[1];
-				pathtracer_bunnyhop_samples[pathtracer_bunnyhop_index].bestangle = bestangle;
-				pathtracer_bunnyhop_samples[pathtracer_bunnyhop_index].pos[0] = sv_player->v.origin[0];
-				pathtracer_bunnyhop_samples[pathtracer_bunnyhop_index].pos[1] = sv_player->v.origin[1];
-				pathtracer_bunnyhop_samples[pathtracer_bunnyhop_index].pos[2] = sv_player->v.origin[2];
-				if (onground0 && !prevonground) {
-					pathtracer_bunnyhop_samples[pathtracer_bunnyhop_index].contact = just_landed;
-				}
-				else {
-					pathtracer_bunnyhop_samples[pathtracer_bunnyhop_index].contact = just_jumped;
-				}
-				pathtracer_bunnyhop_index++;
-				if (pathtracer_bunnyhop_index >= 100)
-					pathtracer_bunnyhop_index = 0;
-			}
-		}
-		if (speed < 1) prevspeed = 0;
-
-		prevprevonground = prevonground;
-		prevonground = onground0;
-	}
-
-	currentclosestbunnyhopDistance = INFINITE;
-	currentclosestbunnyhopIndex = -1;
-	for (int i = 0;i < 100;i++) {
-		pathtracer_bunnyhop_t* pbs = &pathtracer_bunnyhop_samples[i];
-		pbs->selected = false;
-		if (pbs->pos[0] == 0.f && pbs->pos[1] == 0.f && pbs->pos[2] == 0.f) {
-			break;
-		}
-		GLfloat startPos[3];
-		startPos[0] = pbs->pos[0];
-		startPos[1] = pbs->pos[1];
-		startPos[2] = pbs->pos[2] - 20.f;
-
-		int dx = sv_player->v.origin[0] - startPos[0];
-		int dy = sv_player->v.origin[1] - startPos[1];
-		int dz = sv_player->v.origin[2] - startPos[2];
-		float distance = dx * dx + dy * dy + dz * dz;
-		if (distance < currentclosestbunnyhopDistance) {
-			currentclosestbunnyhopDistance = distance;
-			currentclosestbunnyhopIndex = i;
-		}
-	}
-
-
-	if (currentclosestbunnyhopIndex >= 0 && currentclosestbunnyhopDistance < 20000.f) {
-		pathtracer_bunnyhop_samples[currentclosestbunnyhopIndex].selected = true;
-
-		int scale = Sbar_GetScaleAmount();
-		int size = Sbar_GetScaledCharacterSize();
-		int x = ((vid.width - (int)(160 * scale)) / 2) + (show_speed_x.value * size);
-		int y = ELEMENT_Y_COORD(show_speed) + 10.f * scale;
-
-		Draw_String(x, y, pathtracer_bunnyhop_samples[currentclosestbunnyhopIndex].str, true);
-	}
-}
-
 /*
 ==============
 SCR_DrawSpeed
